requirements.txt
Django==4.2.1
Faker==18.9.0
Pillow==9.5.0
crispy-bootstrap5==0.7
django-countries==7.5.1
django-crispy-forms==2.0
django_filter==23.2
factory_boy==3.2.1
gunicorn==20.1.0
martor==1.6.26